+++ src/components/layout/content/login/LoginErrorMessage.js
@@ -1,0 +1,9 @@
+import "../../../../styles/Login.css";
+
+const LoginErrorMessage = ({ message }) => {
+  return (
+    <div className={`errorMessage ${message ? "active" : ""}`}>{message}</div>
+  );
+};
+
+export default LoginErrorMessage;
