+++ src/components/layout/content/login/LoginMain.js
@@ -1,0 +1,15 @@
+import React from "react";
+import LoginTitle from "./LoginTitle";
+import "../../../../styles/Login.css";
+const LoginMain = ({ children }) => {
+    return (
+        <div className="loginPage">
+            <div className="loginContainer">
+                <LoginTitle />
+                {children}
+            </div>
+        </div>
+    );
+};
+
+export default LoginMain;
