--- src/pages/login/ErrorMessage.js
@@ -1,7 +1,0 @@
-const ErrorMessage = ({ message }) => {
-  return (
-    <div className={`errorMessage ${message ? "active" : ""}`}>{message}</div>
-  );
-};
-
-export default ErrorMessage;
