--- src/pages/login/Header.js
@@ -1,12 +1,0 @@
-import SlashLogo from "../../assets/images/logo.png";
-
-const Header = () => {
-  return (
-    <div className="loginHeader">
-      <img src={SlashLogo} alt="slash" />
-      <h3>SLASH</h3>
-    </div>
-  );
-};
-
-export default Header;
